pyproject.toml
[project]
name = "llm-ktr-lineage"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "loguru>=0.7.3",
    "pydantic-ai>=1.0.13",
    "python-dotenv>=1.1.1",
    "tqdm>=4.67.1",
]
